+++ src/App.js
@@ -36,6 +36,18 @@ class App extends Component {
 
 
 
+  componentDidMount() {
+
+    /**
+     * Бывает, не сразу перерендеривается после получения схемы,
+     * поэтому обновляем компонент
+     */
+    setTimeout(() => this.forceUpdate(), 1000);
+
+    super.componentDidMount && super.componentDidMount();
+  }
+
+
   /**
    * Устанавливаем новые фильтры
    */